+++ deploySamples.sh
@@ -50,10 +50,18 @@ mvn clean install -P samples || { echo "Maven build failed" ; exit 1; }
 
 
 
-# GWT sample
-box "Build GWT sample"
-cd samples/gwt
-mvn install -P prod || { echo "Maven build failed" ; exit 1; }
+# GWT 2.8 sample
+box "Build GWT 2.8 sample"
+cd samples/gwt28
+mvn package -P prod || { echo "Maven build failed" ; exit 1; }
+cd $ROOT
+
+
+
+# GWT 3.0 sample
+box "Build GWT 3.0 sample"
+cd samples/gwt30
+mvn package -P prod || { echo "Maven build failed" ; exit 1; }
 cd $ROOT
 
 
@@ -61,7 +69,15 @@ cd $ROOT
 # J2CL sample
 box "Build J2CL sample"
 cd samples/j2cl
-mvn install -P prod || { echo "Maven build failed" ; exit 1; }
+mvn package -P prod || { echo "Maven build failed" ; exit 1; }
+cd $ROOT
+
+
+
+# Crysknife sample
+box "Build Crysknife sample"
+cd samples/crysknife
+mvn package -P prod || { echo "Maven build failed" ; exit 1; }
 cd $ROOT
 
 
@@ -72,8 +88,10 @@ rm -rf /tmp/elemento
 cd /tmp/
 git clone -b gh-pages --single-branch git@github.com:hal/elemento.git
 cd elemento
-cp -R $ROOT/samples/gwt/target/sample-gwt-*/todo gwt
-cp -R $ROOT/samples/j2cl/target/classes/META-INF/resources j2cl
+cp -R $ROOT/samples/gwt28/target/sample-gwt28/todo gwt28
+cp -R $ROOT/samples/gwt30/target/sample-gwt30/todo gwt30
+cp -R $ROOT/samples/j2cl/target/sample-j2cl j2cl
+cp -R $ROOT/samples/crysknife/target/sample-crysknife crysknife
 date > .build
 git add --all
 git commit -am "Update samples"