+++ samples/gwt30/src/main/java/org/jboss/elemento/sample/gwt/client/Todo.java
@@ -13,15 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.elemento.sample.crysknife;
+package org.jboss.elemento.sample.gwt.client;
 
-import javax.inject.Singleton;
+import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsType;
 
-import elemental2.dom.HTMLElement;
-import org.jboss.gwt.elemento.core.IsElement;
-import org.treblereel.gwt.crysknife.annotation.Templated;
+@JsType(isNative = true, namespace = JsPackage.GLOBAL, name = "Object")
+public class Todo {
 
-@Singleton
-@Templated("Todo.html#info")
-public class FooterElement implements IsElement<HTMLElement> {
+    public String id;
+
+    public String text;
+
+    public boolean completed;
 }